News_Article_Summarizer/requirements.txt
groq
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
ipywidgets